Use more reliable way to check linger status
Before Ubuntu 18, loginctl user-status would not indicate whether linger
was enabled. As a result, it was easy to deploy an app via tomo and
forget to turn linger on, sidekiq would fail to stay running, and this
was hard to troubleshoot.

We can make this more reliable by checking the filesystem for linger
status, since what loginctl enable-linger does under the hood is just
create a file under /var/lib/systemd/linger/. Now tomo:setup_systemd
will provide a helpful error message if linger is not enabled, even on
older versions of Ubuntu.